tests for parallel ordered loop moved into separate file
