+++ src/core/tests/test_parallel_for.cc
@@ -89,80 +89,6 @@ TEST(parallel, for_dynamic_nested) {
 
 
 
-static void test_parallel_for_ordered(size_t n)
-{
-  std::atomic_flag global = ATOMIC_FLAG_INIT;
-  std::vector<int> done(n, 0);
-  std::atomic<size_t> next_ordered {0};
-
-  class OTask : public dt::OrderedTask {
-    private:
-      std::atomic_flag& executing_global_;
-      std::vector<int>& done_;
-      std::atomic<size_t>& next_ordered_;
-      std::atomic_flag executing_local_ = ATOMIC_FLAG_INIT;
-      size_t : 56;
-
-    public:
-      OTask(std::atomic_flag& global,
-            std::atomic<size_t>& next_ordered, std::vector<int>& done)
-        : executing_global_(global),
-          done_(done),
-          next_ordered_(next_ordered) {}
-
-      void start(size_t j) override {
-        ASSERT_FALSE(executing_local_.test_and_set());
-        ASSERT_LT(j, done_.size());
-        ASSERT_EQ(done_[j], 0);
-        done_[j] = 1;
-        executing_local_.clear();
-      }
-
-      void order(size_t j) override {
-        ASSERT_FALSE(executing_global_.test_and_set());
-        ASSERT_FALSE(executing_local_.test_and_set());
-        ASSERT_EQ(next_ordered_.load(), j);
-        next_ordered_++;
-        ASSERT_EQ(done_[j], 1);
-        done_[j] = 2;
-        executing_local_.clear();
-        executing_global_.clear();
-      }
-
-      void finish(size_t j) override {
-        ASSERT_FALSE(executing_local_.test_and_set());
-        ASSERT_EQ(done_[j], 2);
-        done_[j] = 3;
-        executing_local_.clear();
-      }
-  };
-
-  dt::parallel_for_ordered(n, dt::NThreads(),
-    [&]{
-      return std::make_unique<OTask>(global, next_ordered, done);
-    });
-
-  ASSERT_EQ(next_ordered.load(), n);
-  for (size_t i = 0; i < n; ++i) {
-    ASSERT_EQ(done[i], 3);
-  }
-}
-
-
-TEST(parallel, for_ordered1) {
-  test_parallel_for_ordered(1723);
-}
-
-TEST(parallel, for_ordered2) {
-  size_t nth0 = dt::num_threads_in_pool();
-  dt::thpool->resize(2);
-  test_parallel_for_ordered(1723);
-  dt::thpool->resize(nth0);
-}
-
-
-
-
 
 }  // namespace dttest
 #endif
